+++ contrib/storage-http/src/test/resources/data/token_refresh.json
@@ -1,0 +1,6 @@
+{
+  "access_token": "token 2.0",
+  "refresh_token": "refresh 2.0",
+  "token_type": "Bearer",
+  "expires": 3600
+}
